refactor crew grouping logic to display multiple jobs under a single person
